remove unneeded try catch code and seperate RblFloat.* to a new spec file
